requirements.txt
azure-storage==0.36.0
blessings==1.7
configparser==4.0.2
